requirements.txt
asgiref==3.9.1
Django==5.2.4
django-jazzmin==3.0.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=336dc47db1955fe7451634a6aacf203887bd98d243e29ceeff7b54213fcf4e5f
mysqlclient==2.2.7
pillow==11.3.0
python-magic-bin==0.4.14
sqlparse==0.5.3
tzdata==2025.2
